Made New Web Page for Website
After finishing the first page of the website, or the homepage, I
created the 2nd web page and made it a replica of the homepage. This is
to give the 2nd web page the same structure as the homepage in the most
efficient manner possible. Also, the 2nd web page is now set up and
ready to modify as needed.